+++ FoodOrderApi/Model/DTO/RestaurantsandMenusDTO.cs
@@ -3,6 +3,6 @@
     public class RestaurantsandMenusDTO
     {
         public List<DisplayMenuDTO> menu { get; set; }
-        public List<DisplayRestaurantDTO> restaurant { get; set; }
+        public List<RestaurantDTO> restaurant { get; set; }
     }
 }
